+++ src/dialogs/linkdialog.cpp
@@ -12,92 +12,80 @@
 
 LinkDialog::LinkDialog(QString dialogTitle, QWidget *parent) :
         MasterDialog(parent),
-    ui(new Ui::LinkDialog)
-{
+        ui(new Ui::LinkDialog) {
     ui->setupUi(this);
 
     this->firstVisibleNoteListRow = 0;
 
-    if ( dialogTitle != "" )
-    {
-        this->setWindowTitle( dialogTitle );
+    if (dialogTitle != "") {
+        this->setWindowTitle(dialogTitle);
     }
 
     QStringList nameList = Note::fetchNoteNames();
     ui->searchLineEdit->installEventFilter(this);
-    ui->notesListWidget->addItems( nameList );
-    ui->notesListWidget->setCurrentRow( 0 );
+    ui->notesListWidget->addItems(nameList);
+    ui->notesListWidget->setCurrentRow(0);
 }
 
-LinkDialog::~LinkDialog()
-{
+LinkDialog::~LinkDialog() {
     delete ui;
 }
 
-void LinkDialog::on_searchLineEdit_textChanged(const QString &arg1)
-{
+void LinkDialog::on_searchLineEdit_textChanged(const QString &arg1) {
     // search notes when at least 2 characters were entered
-    if ( arg1.count() >= 2 )
-    {
-        QList<QString> noteNameList = Note::searchAsNameList( arg1, true );
+    if (arg1.count() >= 2) {
+        QList<QString> noteNameList = Note::searchAsNameList(arg1, true);
         this->firstVisibleNoteListRow = -1;
 
-        for(int i = 0; i < this->ui->notesListWidget->count(); ++i)
-        {
-            QListWidgetItem* item = this->ui->notesListWidget->item(i);
-            if ( noteNameList.indexOf( item->text() ) < 0 )
-            {
-                item->setHidden( true );
-            }
-            else
-            {
-                if ( this->firstVisibleNoteListRow < 0 ) this->firstVisibleNoteListRow = i;
-                item->setHidden( false );
+        for (int i = 0; i < this->ui->notesListWidget->count(); ++i) {
+            QListWidgetItem *item = this->ui->notesListWidget->item(i);
+            if (noteNameList.indexOf(item->text()) < 0) {
+                item->setHidden(true);
+            } else {
+                if (this->firstVisibleNoteListRow < 0) {
+                    this->firstVisibleNoteListRow = i;
+                }
+                item->setHidden(false);
             }
         }
-    }
-    // show all items otherwise
-    else
-    {
+    } else {  // show all items otherwise
         this->firstVisibleNoteListRow = 0;
 
-        for(int i = 0; i < this->ui->notesListWidget->count(); ++i)
-        {
-            QListWidgetItem* item = this->ui->notesListWidget->item(i);
-            item->setHidden( false );
+        for (int i = 0; i < this->ui->notesListWidget->count(); ++i) {
+            QListWidgetItem *item = this->ui->notesListWidget->item(i);
+            item->setHidden(false);
         }
     }
 }
 
-QString LinkDialog::getSelectedNoteName()
-{
-    return ui->notesListWidget->currentRow() > -1 ? ui->notesListWidget->currentItem()->text() : "";
+QString LinkDialog::getSelectedNoteName() {
+    return ui->notesListWidget->currentRow() > -1
+           ? ui->notesListWidget->currentItem()->text() : "";
 }
 
-QString LinkDialog::getURL()
-{
+QString LinkDialog::getURL() {
     return ui->urlEdit->text();
 }
 
 //
 // Event filters on the NoteSearchDialog
 //
-bool LinkDialog::eventFilter(QObject* obj, QEvent *event)
-{
-    if ( obj == ui->searchLineEdit )
-    {
-        if ( event->type() == QEvent::KeyPress )
-        {
+bool LinkDialog::eventFilter(QObject *obj, QEvent *event) {
+    if (obj == ui->searchLineEdit) {
+        if (event->type() == QEvent::KeyPress) {
             QKeyEvent *keyEvent = static_cast<QKeyEvent *>(event);
 
-            // set focus to the notes list if Key_Down or Key_Tab were pressed in the search line edit
-            if ( ( keyEvent->key() == Qt::Key_Down ) || ( keyEvent->key() == Qt::Key_Tab ) )
-            {
+            // set focus to the notes list if Key_Down or Key_Tab were pressed
+            // in the search line edit
+            if ((keyEvent->key() == Qt::Key_Down) ||
+                (keyEvent->key() == Qt::Key_Tab)) {
                 // choose an other selected item if current item is invisible
                 QListWidgetItem *item = ui->notesListWidget->currentItem();
-                if ( ( item != NULL ) && ui->notesListWidget->currentItem()->isHidden() && ( this->firstVisibleNoteListRow >= 0 ) )
-                {
-                    ui->notesListWidget->setCurrentRow( this->firstVisibleNoteListRow );
+                if ((item != NULL) &&
+                    ui->notesListWidget->currentItem()->isHidden() &&
+                    (this->firstVisibleNoteListRow >= 0)) {
+                    ui->notesListWidget->setCurrentRow(
+                            this->firstVisibleNoteListRow);
                 }
 
                 // give the keyboard focus to the notes list widget
@@ -106,17 +94,15 @@ bool LinkDialog::eventFilter(QObject* obj, QEvent *event)
             }
         }
         return false;
-    }
-    else if ( obj == ui->notesListWidget )
-    {
-        if ( event->type() == QEvent::KeyPress )
-        {
+    } else if (obj == ui->notesListWidget) {
+        if (event->type() == QEvent::KeyPress) {
             QKeyEvent *keyEvent = static_cast<QKeyEvent *>(event);
 
-            // set focus to the note text edit if Key_Return or Key_Tab were pressed in the notes list
-            if ( ( keyEvent->key() == Qt::Key_Return ) || ( keyEvent->key() == Qt::Key_Tab ) )
-            {
-                //focusNoteTextEdit();
+            // set focus to the note text edit if Key_Return or Key_Tab
+            // were pressed in the notes list
+            if ((keyEvent->key() == Qt::Key_Return) ||
+                (keyEvent->key() == Qt::Key_Tab)) {
+                // focusNoteTextEdit();
                 return true;
             }
         }
@@ -126,11 +112,10 @@ bool LinkDialog::eventFilter(QObject* obj, QEvent *event)
     return LinkDialog::eventFilter(obj, event);
 }
 
-void LinkDialog::on_notesListWidget_doubleClicked(const QModelIndex &index)
-{
-    Q_UNUSED( index );
+void LinkDialog::on_notesListWidget_doubleClicked(const QModelIndex &index) {
+    Q_UNUSED(index);
     this->close();
-    this->setResult( QDialog::Accepted );
+    this->setResult(QDialog::Accepted);
 }
 
 extern "C" size_t decode_html_entities_utf8(char *dest, const char *src);
@@ -140,39 +125,43 @@ extern "C" size_t decode_html_entities_utf8(char *dest, const char *src);
  * @param url
  * @return
  */
-QString LinkDialog::getTitleForUrl( QUrl url )
-{
+QString LinkDialog::getTitleForUrl(QUrl url) {
     QNetworkAccessManager *manager = new QNetworkAccessManager(this);
     QEventLoop loop;
     QTimer timer;
 
     timer.setSingleShot(true);
     connect(&timer, SIGNAL(timeout()), &loop, SLOT(quit()));
-    connect(manager, SIGNAL(finished(QNetworkReply*)), &loop, SLOT(quit()));
+    connect(manager, SIGNAL(finished(QNetworkReply *)), &loop, SLOT(quit()));
 
     // 5 sec timeout for the request
     timer.start(5000);
 
-    QNetworkReply *reply = manager->get( QNetworkRequest( url ) );
+    QNetworkReply *reply = manager->get(QNetworkRequest(url));
     loop.exec();
 
     // if we didn't get a timeout let's fetch the title of the webpage
-    if ( timer.isActive() )
-    {
+    if (timer.isActive()) {
         // get the text from the network reply
         QString html = reply->readAll();
 
         // parse title from webpage
-        QRegularExpression regex( "<title>(.*)</title>" );
-        QRegularExpressionMatch match = regex.match( html );
+        QRegularExpression regex("<title>(.*)</title>");
+        QRegularExpressionMatch match = regex.match(html);
         QString title = match.captured(1);
 
         // decode HTML entities
         HTMLEntities htmlEntities;
-        title = htmlEntities.decodeHtmlEntities( title );
+        title = htmlEntities.decodeHtmlEntities(title);
 
         // replace some other characters we don't want
-        title.replace( "[", "(" ).replace( "]", ")" ).replace( "<", "(" ).replace( ">", ")" );
+        title.replace("[", "(")
+                .replace("]", ")")
+                .replace("<", "(")
+                .replace(">", ")")
+                // this character used by GitHub disorients the markup
+                // highlighter
+                .replace("·", "-");
 
         // trim whitespaces and return title
         return title.simplified();
@@ -185,13 +174,12 @@ QString LinkDialog::getTitleForUrl( QUrl url )
 /**
  * @brief Selects a local file to link to
  */
-void LinkDialog::on_fileUrlButton_clicked()
-{
-    QUrl fileUrl = QFileDialog::getOpenFileUrl( this, tr( "Select file to link to" ) );
+void LinkDialog::on_fileUrlButton_clicked() {
+    QUrl fileUrl = QFileDialog::getOpenFileUrl(this,
+                                               tr("Select file to link to"));
     QString fileUrlString = fileUrl.toString();
 
-    if ( fileUrlString != "" )
-    {
-        ui->urlEdit->setText( fileUrlString );
+    if (fileUrlString != "") {
+        ui->urlEdit->setText(fileUrlString);
     }
 }